Enum pci_types::device_type::DeviceType

pub enum DeviceType {

    Unknown,
    LegacyVgaCompatible,
    LegacyNotVgaCompatible,
    ScsiBusController,
    IdeController,
    FloppyController,
    IpiBusController,
    RaidController,
    AtaController,
    SataController,
    SasController,
    OtherMassStorageController,
    EthernetController,
    TokenRingController,
    FddiController,
    AtmController,
    IsdnController,
    WorldFipController,
    PicmgController,
    OtherNetworkController,
    VgaCompatibleController,
    XgaController,
    ThreeDController,
    OtherDisplayController,
    VideoDevice,
    AudioDevice,
    TelephonyDevice,
    OtherMultimediaDevice,
    RamController,
    FlashController,
    OtherMemoryController,
    HostBridge,
    IsaBridge,
    EisaBridge,
    McaBridge,
    PciPciBridge,
    PcmciaBridge,
    NuBusBridge,
    CardBusBridge,
    RacewayBridge,
    SemiTransparentPciPciBridge,
    InfinibandPciHostBridge,
    OtherBridgeDevice,
    SerialController,
    ParallelPort,
    MultiportSerialController,
    Modem,
    GpibController,
    SmartCard,
    OtherCommunicationsDevice,
    InterruptController,
    DmaController,
    SystemTimer,
    RtcController,
    GenericPciHotPlugController,
    SdHostController,
    OtherSystemPeripheral,
    KeyboardController,
    Digitizer,
    MouseController,
    ScannerController,
    GameportController,
    OtherInputController,
    GenericDockingStation,
    OtherDockingStation,
    Processor386,
    Processor486,
    ProcessorPentium,
    ProcessorAlpha,
    ProcessorPowerPc,
    ProcessorMips,
    CoProcessor,
    FirewireController,
    AccessBusController,
    SsaBusController,
    UsbController,
    FibreChannelController,
    SmBusController,
    InfiniBandController,
    IpmiController,
    SercosController,
    CanBusController,
    IrdaController,
    ConsumerIrController,
    RfController,
    BluetoothController,
    BroadbandController,
    Ethernet5GHzController,
    Ethernet24GHzController,
    OtherWirelessController,
    IntelligentIoController,
    TvSatelliteCommunicationsController,
    AudioSatelliteCommunicationsController,
    VoiceSatelliteCommunicationsController,
    DataSatelliteCommunicationsController,
    NetworkCryptionController,
    EntertainmentCryptionController,
    OtherCryptionController,
    DpioModule,
    PerformanceCounter,
    CommunicationsSynchronizationController,
    ManagementCard,
    OtherSignalProcessingController,

}

Combines the Base Class and the Sub-class of a device to classify it into a DeviceType. Combined with the device’s Interface, this can be enough to know how to drive the device.

Variants

Unknown

LegacyVgaCompatible

LegacyNotVgaCompatible

ScsiBusController

IdeController

FloppyController

IpiBusController

RaidController

AtaController

SataController

SasController

OtherMassStorageController

EthernetController

TokenRingController

FddiController

AtmController

IsdnController

WorldFipController

PicmgController

OtherNetworkController

VgaCompatibleController

XgaController

ThreeDController

OtherDisplayController

VideoDevice

AudioDevice

TelephonyDevice

OtherMultimediaDevice

RamController

FlashController

OtherMemoryController

HostBridge

IsaBridge

EisaBridge

McaBridge

PciPciBridge

PcmciaBridge

NuBusBridge

CardBusBridge

RacewayBridge

SemiTransparentPciPciBridge

InfinibandPciHostBridge

OtherBridgeDevice

SerialController

ParallelPort

MultiportSerialController

Modem

GpibController

SmartCard

OtherCommunicationsDevice

InterruptController

DmaController

SystemTimer

RtcController

GenericPciHotPlugController

SdHostController

OtherSystemPeripheral

KeyboardController

Digitizer

MouseController

ScannerController

GameportController

OtherInputController

GenericDockingStation

OtherDockingStation

Processor386

Processor486

ProcessorPentium

ProcessorAlpha

ProcessorPowerPc

ProcessorMips

CoProcessor

FirewireController

AccessBusController

SsaBusController

UsbController

FibreChannelController

SmBusController

InfiniBandController

IpmiController

SercosController

CanBusController

IrdaController

ConsumerIrController

RfController

BluetoothController

BroadbandController

Ethernet5GHzController

Ethernet24GHzController

OtherWirelessController

IntelligentIoController

TvSatelliteCommunicationsController

AudioSatelliteCommunicationsController

VoiceSatelliteCommunicationsController

DataSatelliteCommunicationsController

NetworkCryptionController

EntertainmentCryptionController

OtherCryptionController

DpioModule

PerformanceCounter

CommunicationsSynchronizationController

ManagementCard

OtherSignalProcessingController