Crate pci_types

Modules

capability

device_type

Structs

CommandRegister

EndpointHeader

Endpoints have a Type-0 header, so the remainder of the header is of the form:

PciAddress

The address of a PCIe function.

PciHeader

Every PCI configuration region starts with a header made up of two parts:

PciPciBridgeHeader

PCI-PCI Bridges have a Type-1 header, so the remainder of the header is of the form:

StatusRegister

Enums

Bar

BarWriteError

DevselTiming

Slowest time that a device will assert DEVSEL# for any bus command except Configuration Space read and writes

HeaderType

Constants

MAX_BARS

Traits

ConfigRegionAccess

Type Aliases

BaseClass

DeviceId

DeviceRevision

Interface

InterruptLine

InterruptPin

SubClass

SubsystemId

SubsystemVendorId

VendorId