Crate pci_types

Modules

device_type

Structs

EndpointHeader

Endpoints have a Type-0 header, so the remainder of the header is of the form:

PciAddress

PCIe supports 65536 segments, each with 256 buses, each with 32 slots, each with 8 possible functions. We cram this into a u32:

PciHeader

Every PCI configuration region starts with a header made up of two parts:

Enums

Bar

Constants

HEADER_TYPE_CARDBUS_BRIDGE

HEADER_TYPE_ENDPOINT

HEADER_TYPE_PCI_PCI_BRIDGE

MAX_BARS

Traits

ConfigRegionAccess

Type Definitions

BaseClass

DeviceId

DeviceRevision

HeaderType

Interface

SubClass

VendorId