Crate pci_types

Modules

capability

device_type

Structs

EndpointHeader

Endpoints have a Type-0 header, so the remainder of the header is of the form:

PciAddress

PCIe supports 65536 segments, each with 256 buses, each with 32 slots, each with 8 possible functions. We cram this into a u32:

PciHeader

Every PCI configuration region starts with a header made up of two parts:

PciPciBridgeHeader

PCI-PCI Bridges have a Type-1 header, so the remainder of the header is of the form:

StatusRegister

Enums

Bar

DevselTiming

Slowest time that a device will assert DEVSEL# for any bus command except Configuration Space read and writes

HeaderType

Constants

MAX_BARS

Traits

ConfigRegionAccess

Type Aliases

BaseClass

DeviceId

DeviceRevision

Interface

InterruptLine

InterruptPin

SubClass

SubsystemId

SubsystemVendorId

VendorId