pci_types

Module capability

Structs

CapabilityIterator

MsiCapability

MsixCapability

PciCapabilityAddress

Enums

MultipleMessageSupport

Specifies how many MSI interrupts one device can have. Device will modify lower bits of interrupt vector to send multiple messages, so interrupt block must be aligned accordingly.

PciCapability

PCI capabilities

TriggerMode

When device should trigger the interrupt